requirements.txt
requests
pandas
python-dotenv
flask
schedule